+++ GameRunner/RenderPass2D.cpp
@@ -167,7 +167,7 @@ void RenderPass2D::BuildRenderPipeline()
                 .pImageInfo = ColorDescriptorImage.data(),
             }
         };
-        Renderer_UpdateDescriptorSet(descriptorSets.data(), descriptorSets.size());
+        Renderer_UpdateDescriptorSet(descriptorSets.data(), static_cast<uint32_t>(descriptorSets.size()));
     }
 
     /*  std::vector<VkVertexInputBindingDescription> bindingDescriptions
